Merge pcaps on Dalton Conroller instead of the Agent.
Suricata can only read a single pcap in pcap read mode
so if multiple pcaps are submitted, then must first be
combined.  Previously this was done on the Agent which
is nice for distributing load if this was heavy laden but
it also add complexity to the Agent and requires it to have
additional programs (mergecap) installed. And given that
wireshark has to be installed (large!) to get mergecap, it
makes sense to just do this on the Controller and keep the
Agents slim and unburdened.