concatenate unified2 data and b64 encode to send back to Controller
Instead of decoding on the agent and sending the text back (would
require u2spewfoo or something like python idstools), just
concatenate the binary files and send them back.  Because we use
urllib2 and not something more featured like Requests, we
have to send the data as www-form-urlencoded which means we
have to base64 to for the transfer which means bloat.
multipart/mixed MIME would be better but at least we won't get URI encoding bloat.
Still need to update Controller code.